imports/refact: Remove envCheckOptionsConsistency entirely
`envCheckOptionConsistency` was *only* modified `False` (via
`withoutOptionsChecking`) when loading primitive modules in
`typeCheckMain`. Those modules do not import anything else, so the
value did not apply transitively to anything.

The value was only used in this Agda.Interaction.Imports in order to
suppress checking for options consistency for the primitive modules. It
did so in two different ways: one with the `miPrimitive` boolean stored
in `ModuleInfo`, the other with this env value.

Now it's simpler.