imports/refact: Remove redundant envCheckOptionConsistency check
`envCheckOptionConsistency` is *only* set to `False` (via
`withoutOptionsChecking`) when loading primitive modules in
`typeCheckMain`. Those modules are also considered builtins.

Therefore the posible cases are:
   False, primitive builtin module → False ∧ ¬True  → False
   True,  non-primitive builtin    → True  ∧ ¬True  → False
   True,  normal module            → True  ∧ ¬False → True

So we only care about if it's a builtin.

And in fact, not addressed here, but the `envCheckOptionConsistency` and
`miPrimitive` flag on `ModuleInfo` represent the same information.