[ code comprehension re #5098 ] rename SourceInfo to Source
"X-info" usually denotes some side information about object X.
In particular, it is not the entirety of X.

The name `SourceInfo` for a record containing the entire source code
both in textual and parsed form plus additional info about it isn't
appropriate imho and every time trips up my mind.

Cf.: If you read

- "Info about Linear Algebra", you would not expect this to contain
the whole lecture series on linear algebra;

- "Info about the Bible", you would not expect this to contain the
whole text of the Bible.

So I renamed `SourceInfo` to just `Source`, because it is THE THING
ITSELF and not just some information ABOUT it.