[ closed #5112 ] Used local `stack-X.Y.Z.yaml` files.
The `stack-X.Y.Z.yaml` files from the `fix-whitespace` program are
used for installing it.

Currently the `Makefile` uses `stack` if the `stack.yaml` file (which
is .gitignored) is in the root directory. So you can run

    ln -s stack-X.Y.Z.yaml stack.yaml

for using `stack` with GHC X.Y.Z

The fix use a similar approach. That is, a `stack.yaml` (also
.gitignored) is required in the root directory of the `fix-whitespace`
program (i.e. `src/fix-whitespace`).

For creating both `stack.yaml` files (for Agda and for the
`fix-whitespace` program) you can run

    make set-default-stack-file GHC_COMPILER=8.10.2