[ fix ] flip the equality for better ergonomics
As demonstrated by `filter-filter` we often want to use the equality
to unfold a computation whose result we already know. This is annoying
if the equality is `var 0 = e` because `e` is the computation we want
to unfold and `var 0` has typically been refined to a pattern.
With the flipped equality `e = var 0` we get a much more useful assumption.