[ fixed #5288 ] treeless de Bruijn problem "withContextSize" (#5311)
For the translation of the catch-all clause of the enclosed testcase,
```agda
f : Nat → Nat → Nat
f _ zero = 0
f m      = λ _ → m
```
we enter from a context of length 2 (`commonArity`) to a context of
length 1 (where `λ _ → m` lives).

The previous code would translate `λ _ → m` in context `0:m` and then
simply drop back to the context `[1:m,0:_]`, resulting in a wrong de
Bruijn index (`λ _ → 1`).  We now keep translating in context `[1:m]`
resulting in `λ _ → 2` which is correct when going back to `[1:m,0:_]`.