[ #5236 ] `constructor` no layout keyword in `record ... where`
`constructor` should be a layout keyword in `interleaved mutual`
blocks, but not in `record`-declarations.

To distinguish these situations, we rember in `LayoutContext` which
keyword introduced each `LayoutBlock`.  When encountering
`KwConstructor`, we search through the blocks whether we recently had
a `mutual` (then we start a layout block) or rather a `where` (then we
do not start a layout block).