[ fixed #5334 ] More powerful sorting of constructors to data in interleaved mutual
@gallais:
Patterns cannot have holes in them and I am reusing the `couldBeCallOf`
logic that is used to group blocks of function clauses inside a single
definition to identify potential parent data types.

@andreasabel:
Looking at `couldBeCallOf`, it actually does not make use of the pattern
structure, only the names a pattern contains.  Thus, we just patch the
translation to pattern to map all non-pattern things to `WildP`.