[ fix #5478 ] properly reuse IP Meta in a record context
An interaction meta might be created in a context with one binding for
each record field, but then reused in a context with just a single
binding of the record variable.

This patch tries to correctly instantiate the original meta (created
during the constructor type construction run) when used during a
second type-checking run (this time, the record declarations).