[ cifm ] Efficient retrieval of milestone (needs unreleased github-0.28)
The latest github-0.27 Haskell API misses parametrization of the "repo
issues" request by a milestone id.

Thus, the only way to get the closed issues for a milestone with
github-0.27 is to download all issues.

By extending the github-0.27 API with
```
    optionsMilestone :: Id Milestone -> IssueRepoMod
```
we can communicate to github that we only want the issues for a
certain milestone.

Instead of waiting for minutes, we now get the closed issues for
e.g. 2.6.3 instantaneously.