[ fix #5807 ] Clarify --irrelevant-projections and irrAx.
Attempt to clarify some of my confusion about the documentation:
* Prop serves as an alternative to irrelevance
* The irrelevance axiom is a form of non-constructive choice
* Irrelevant projections are justified by the irrelevance axiom

All of these things can already be inferred from what is already
written, but this states them explicitly. In hindsight it's
obvious, but it took me a while to figure it out. Hopefully
this is helpful and I'm not just bad at reading...?

I did not mention the reason that irrelevant projections
cannot be reduced because that doesn't seem like the sort of
thing included elsewhere in the manual. (The reason being
that it's inconsistent.)

I also think it would also be helpful to mention what options are
unsafe when they are introduced, but that's outside the scope
of this commit.