Fix #5823: checking for loops when constructing singleton inhabitant
For inductive record types with eta-equality, the check for singleton
type (constructing the unique inhabitant) was looping for
non-terminating records.

As a remedy, the check now keeps a record of already unfolded recursive
record types.  This alone is too restrictive, because "terminating"
record types should always be unfolded (to their field telescope).
The Data.Record module from the standard-library relied on this
feature in order to construct the record signature automatically by
basically just eta-expansion.

To keep Data.Record etc working, this commit also adds a termination
checker for record types.  To this end, a record type

    record R pars : Set where
      field
        f1 : A1
        ...
        fn : An

is considered as a list of definitions

    R pars = A1
    ...
    R pars = An

which are subjected to call extraction and termination checking.

Record types that fail the termination check are remembered as such,
setting

    recTerminates = Just False

in the `Record` `Definition`, but no termination error is reported.