Fix #5891: Postponable check that data isn't IUniv, SizeUniv,… (#5892)
Introduces constraint `CheckDataSort` to check that the data and record types 
are not constructed in sorts that do not admit such constructions, 
e.g. IUniv, SizeUniv etc.

Also, fix the `SizeUniv` builtin so that it does not declare `SizeUniv : SizeUniv`.  
This fixes the inconsistency reported in issue #5891.

AIM XXXV codesprint.