Re #4725 #5953: termination: consider dotted variable and record patterns
With #4606, --cubical no longer considered dotted patterns as patterns
for termination certification.  This patch considers dotted variables
and dotted record patterns as benign and considers then as pattern
always in the termination checker, even in the first phase (no
dot-patterns) which is the only phase with --cubical.  (The other
modes have the dot-pattern termination phase as well.)