[ #2753, #4653 ] Changes to the generation of symbolic identifiers.
See the changelog entry for details about the changes.

I added new components to the state (stPreAccess) and
environment (envAccess) instead of reusing the scope, along with
possibly duplicated logic (which might not entirely match the logic of
the scope checker). I suggest that this commit should not be merged.