Use a hashmap lookup for decoding warning names
The original implementation performs a list append, meaning it performs
multiple allocations when parsing warning names. As interface files use
this function to decode the enabled warnings for the original file, it
ends up taking about 15% of decoding time.

Switching over to a hashmap both reduces allocations (from 23Gb to 10Gb)
and runtime (12.3s to 9.5s) when loading interfaces for a test project.