Deserialisation: use array of hashmaps (#5980)
* Use an array for memoising deserialised values

We now use an array of hashmaps, rather than a single hashmap of
(idx, type). As most slots will be a single entry, this makes lookups
faster, resulting in a 10% speedup.

* Deserialise directly to an array

Not actually, as Binary doesn't allow reading to an array. However, this
saves an extra call to List.genericLength and appears to specialise
better, saving another 10%.

* Deserialise arrays in-place

This uses STT to allocate a mutable array and write to it while
deserialising, instead of allocating a separate array. This provides a
small (5-6%) improvement to deserialisation times.

* Fix wrong serialise/deserialise in comment

* Explain why we're not using for_ here.