Do not construct a full map when finding a single name (#6014)
When pretty printing lots of types or terms, we end up doing a large number of name lookups.

The current implementation calls `allNamesInScope`, which allocates a map of names to variables. As we're only looking up a single name, we can avoid this additional work, and only perform it on the variable we're searching for.

In my test project, this reduces the time taken in `Scoping.InverseScopeLookup` from 4.5s to 0.7s.