[ refine #5672 (fix for #5600) ]
Only `alwaysUnblock` when comparing `MissingClauses f` neutrals for a
function `f` of the current mutual block.

This was suggested by @Saizan in
https://github.com/agda/agda/issues/5600#issuecomment-964052741
but the implementation in #5672 did not include the check of whether
the missing clauses are from a function of the current mutual block
(and hence have hope to be filled in still...).

Unfortunately, this refinement does not fix #6008.