A potential optimisation (Θ(n²) to Θ(n log n)).
The expressions refer to the number of equality tests/comparisons
performed for elements of lists of length n.

The lists (lists of type [PlusLevel]) are presumably typically very
short, so in practice this "optimisation" might have little effect. A
quick test (without the use of -foptimise-heavily) suggests that this
is the case for the standard library up to and including Data.Nat.
However, the new code is arguably more readable than the old code, so
I still think it makes sense to perform this change.