[ #6097, #6100 ] Replaced Pi with Tel.
And added a pattern synonym for Pi.

Limited testing (without -foptimise-heavily) suggests that this change
makes Agda a little slower for the standard library up to and
including Data.Nat.

Some forms of code are now handled much more quickly. However, the
problem has not been completely fixed. For instance, limited testing
suggests that Agda is quadratic in the number of arguments of F in
test/Succeed/Issue6097-6.agda.

One remaining problem is perhaps that underAbstractions is (at least)
linear in the size of the telescope. However, I suspect that the main
problem (for test/Succeed/Issue6097-6.agda) is that, for a type like
"(_ … _ : Set₁) → ?", the total size of the Sort arguments to the El
constructors is quadratic in the number of arguments.