Fix #6066: `pattern` with eta record: document; warn
The warning "useless pattern directive" is now extended to records
with inferred eta (was only for declared eta).

However, to make it correct, we have to remember whether the record
constructor was subject to matching before we inferred eta for the
record.
(Note: the positivity checker runs at the end of the mutual block.)

I wonder whether we should rather advice the user to change `pattern`
to `eta-equality` in this case.  This would also save us the
bookkeeping of record constructor matches.