[ #6146 ] Tried to make TopLevelModuleName more efficient.
The type now contains a hash of the module name. This hash is used for
comparisons. When a TopLevelModuleName is created it is checked that
the hash does not collide with previously seen hashes.

The type RawTopLevelModuleName (without the hash) was added. A mapping
between RawTopLevelModuleNames and ModuleNameHashes is maintained in
the state.

The compiler backend interface (Backend') now uses TopLevelModuleName
instead of ModuleName.