Fix #6165 by not insisting on `builtinSet` in the scope checker
Even in the presence of `{-# BUILTIN TYPE Type #-}`, this builtin
might not be defined in the scope checker, as this pragma is fully
processed only in the type checker.  Thus, don't insist on
`builtinSet` in the scope checker, rather produce different dummy
expression for the target of the record constructor.