[ #4743 ] Replaced envModality, removed some instances.
The field envModality was removed in favour of envRelevance and
envQuantity, because (according to some comment) cohesion shouldn't be
in the environment. The LensModality instance for TCEnv and the lens
eModality were also removed.

Furthermore the LensRelevance and LensQuantity instances for TCEnv
were removed. Now every change of the environment's relevance
component is performed using the lens eRelevance, and every change of
the environment's quantity component is performed using eQuantity.