[ #5201 ] Builtin syntax for Σ-types.
Now .agda-lib files are read before Agda files are parsed, so that
options that affect the parser can be read from .agda-lib files.
Before a file has been parsed the top-level module name is not known,
so the search for .agda-lib files starts in the directory containing
the Agda file. If it later turns out that any .agda-lib files that
were found were located too far down the directory hierarchy, then an
error is raised. However, it could be the case that the Agda file
fails to parse because an incorrect .agda-lib file is used, and in
that case the error message could be confusing.

Note that if a symbol called Σ is defined in a module that does not
use --unicode-sigma, and this symbol is printed when a module that
uses --unicode-sigma is type-checked, then the resulting output might
not be syntactically correct (if --unicode-sigma is active).

TODO:
* The standard library and cubical test cases fail:
  * Both use Agda.Builtin.Σ, which has been renamed to Σ̂.
  * The cubical library uses the name sigma, which is now reserved.
* The changelog and user manual have not been edited.
* It might make sense to use the builtin syntax when pretty-printing.