Refactor for #6337: simplify checkOpts and unsafePragmaOptions
`checkOpts` is just excluding `--vim` and `--only-scope-checking`, it does
not modify the options record any longer (maybe did so in the past).
Thus, it's type can be simply `m ()` (for a `MonadError m`).

`unsafePragmaOptions` took `CommandLineOptions` as arguments but ignored
it, so I zapped it.  This allowed to simplify code elsewhere.