Fix #6337: infrastructure for renamed options and option warnings
It is more user-friendly to not flatly rename a commandline or pragma
option, but to keep the old name around with a deprecation warning.

This commit sets up the necessary infrasturcture and applies it to the
renaming from --experimental-lossy-unification to --lossy-unification.

Also adds the first testcase for the --lossy-unification feature.