agda-mode: Further simplify process filter
This change builds on c7492c00016d586ca1426508665bca56a433a6f0,
dropping the need for any queue (of string or a buffer), by inserting
the process output directly into the process buffer and operating on
thereon.  We avoid the overhead of consing lists and strings, while
simplifying the logic by relying more on Lisp primitives.