Generalize WithDefault from Bool to Boolean
Unfortunatly, the default setting must still be 'True' or 'False'
rather that the 'true' or 'false' of the 'Boolean' class,
since this class does not lift naively to the type-level.

Maybe we can figure out to write e.g.

    WithDefault' UnicodeOrAscii 'UnicodeOk

or even

    WithDefault 'UnicodeOk

instead of now

    WithDefault' UnicodeOrAscii 'True

Main obstacle is that we have to generalize `KnownBool (b :: Bool)` to
`KnownBoolean a (b :: a)` but that would require extra boilerplate,
like duplicating `Agda.Utils.TypeLits` for each 'Boolean' type.

Ideally, one would state the iso between on `Bool` and `a` (where
`IsBool a`) only once in the `IsBool` instance, not again on the type
level...

Left for future work / needs research.