Fix inferFunSort bug from #6505
Fixes the following problem:

`HasPTSRule` takes the domain and the sort of the codomain.
It looks like here it was invoked with the _sort_ of the domain rather than the domain itself.
In the test case this lead a constraint `HasPTSRule LockUniv LockUniv` which fails on `funSort Set1 LockUniv`.

As `inferFunSort` adds a `HasPTSRule` constraint, it needs to get the domain as argument not just its sort.