Remove unused imports in 119 modules
I kept redundant imports related to Monoid/Semigroup, so that going
back to GHC 8.2 would be easier.  Not sure whether this was necessary.

As I want e.g.

    import Data.Set (Set)
    import Data.Set qualified as Set

to go together, I don't remove redundant imports of either if the
other one is not redundant.

Some modules now have no unused imports; for these, I kept the warning on:

    {-# OPTIONS_GHC -Wunused-imports #-}

Someone working on a module might find it annoying, so welcome to
switch it off or temporarily disable.