Allow entries `name = path` in the `executables` file.
This is mainly motivated by the testsuite where

    make AGDA_BIN=agda-quicker succeed

would error because `execTC` wants to invoke `agda` but the
`executables` file will naturally allow invocation of `agda-quicker`.

A workaround would be to use symbolic links.
However, since Agda anyway constructs a `name ↦ path` map,
it feels natural to allow the user to specify this map explicitly.