Rewrite highlight-hover.js from scratch
The new implementation is noticeably faster for large `.html` files as it
does not per event iterate through /all/ 'a' HTML elements on a page, but
only those it needs to modify.

This is achieved by the use of the data structures natural to the
problem.  We organize the 'a' elements in a map from their 'href' to
the set of all 'a' elements that have the same 'href'.

Closes #6755 (again) by removing the alien license for the old `highlight-hover.js`.