Fix search on readthedocs.org (#6821)
This addresses a regression in `sphinx_rtd_theme` that causes `sphinxcontrib.jquery` to be not loaded.
Consequently, the search page seems to be hanging; the cause is that `jQuery` is undefined.
Adding `sphinxcontrib.jquery` is a suggested workaround at https://github.com/readthedocs/sphinx_rtd_theme/issues/1452.

Since I see another problem with search locally and in the PR build on rtd.org, I bump to the latest `Sphinx` and `sphinx_rtd_theme`.  This doesn't fix the problem, though.  However, rtd.org seems to be fine with these requirements.
The problem I have been seeing is a `undefined` sprinkled in the `requestUrl` variable in Sphinx's `searchtools.js` which causes page loads to fail for previewing the search hits.  Maybe this problem will disappear in the regular build on rtd.org.

Squashed commits:

* Fix search on readthedocs.org by bumping sphinx_rtd_theme to 1.2.2

1.2.2 includes PR 1448 which fixes issue 1452

- https://github.com/readthedocs/sphinx_rtd_theme/pull/1448
- https://github.com/readthedocs/sphinx_rtd_theme/issues/1452

* Fixup: remove `html_theme_path`

Suggested in https://github.com/readthedocs/sphinx_rtd_theme/issues/1452#issuecomment-1575476244

* Fixup 2: or, add 'sphinxcontrib.jquery' to 'extensions'

* Fixup 3: bump to latest Sphinx and rtd theme