Relax dependency bound to `split < 0.3`
I am not sure what the reason was to have an upper bound that is not a
major version (we had first `< 0.2.3.4` and then `< 0.2.4`).

NB: The PVP recommends a minor upper bound (`< 0.2.4`) only in case that
we define an orphan instance for a type or class defined in the dependency.

Triggered by this issue:
- https://github.com/commercialhaskell/stackage/issues/7141