toggle debug reporting via CPP
Disable generating any code for reporting functions in Agda.TypeChecking.Monad.Debug
if the DEBUG flag is not enabled. This yields noticable build time and runtime
speedups.

Previously the DEBUG flag only had minor use in serialization and parsing.
Those usages are now relegated to the new DEBUG_SERIALISATION and DEBUG_PARSING
flags.

There are two new functions, `alwaysReportSLn` and `alwaysReportSDoc`, which are
always active, regardless of the DEBUG flag. These are used wherever we
previously printed with verbosity 1, so default printing output is unchanged.