toggle debug reporting via CPP (#6863)
* toggle debug reporting via CPP

Disable generating any code for reporting functions in `Agda.TypeChecking.Monad.Debug`
if the DEBUG flag is not enabled. This yields noticable build time and runtime
speedups.

Previously the DEBUG flag only had minor use in serialization and parsing.
Those usages are now relegated to the new DEBUG_SERIALISATION and DEBUG_PARSING
flags.

There are two new functions, `alwaysReportSLn` and `alwaysReportSDoc`, which are
always active, regardless of the DEBUG flag. These are used wherever we
previously printed with verbosity 1, so default printing output is unchanged.

Changes in the Makefile and CI:

- `make install-bin` builds without debug output.
- There is a new `install-bin-debug` command which enables debug on the top of `install-bin`. 
- The `quicker-`, `fast-` and `v1-install` commands build with debug output.
- The `test.yml` workflow builds Agda with `install-bin-debug` because some tests rely on
  verbose printing.