Rename record constructors in ScopeM
Since we have a new name for the record, we have to make sure that the
new record is also associated with a constructor.

But we also can't associate it with the old constructor, since that
won't be applied to any parameters.

So we rename the constructor when renaming the record, and fortunately
the type checker is smart enough to recognise that it doesn't have to
come up with a *third* name for the constructor, so the parameters are
correctly instantiated.