Provide a `.agda-lib` for Agda builtins (#6988)
* Ignore directories with an .agda-lib suffix

Co-authored-by: ibbem <ibbem@ibbem.net>

* Fix #4526, #4613: Add .agda-lib for builtins

By default Agda tries to traverse the filesystem upwards until it hits
the filesystem root or it finds a .agda-lib file. This traversal can be
very slow if the Agda builtin sources are contained in a big directory
(e.g. on NixOS where /nix/store, inside which Agda is stored, often
contains ten to hundred thousand files). Hence we introduce the
.agda-lib file to prevent this traversal for the Agda builtins.

Co-authored-by: Guillaume Allais <guillaume.allais@ens-lyon.org>

* Decide between interface files based on their existence

---------

Co-authored-by: Guillaume Allais <guillaume.allais@ens-lyon.org>