agda-mode: Improve robustness of the Emacs' output filter
Since the process output can be chunked inconveniently (e.g. breaking
mid-line), we need to make sure that we can recognise and handle these
situations.  This includes making sure we don't mis-`read' part of a
command, then run into a type error later on, but also restoring the
previous state, so that we can hope to correctly parse the data upon
receiving the next output chunk.