Refactor: newtype KwRange for Ranges that just span keywords
Motivation: in concrete syntax for blocks, we can either store the
range of the block keyword or the range of the whole block.

To not confuse these, we introduce here a newtype `KwRange` for the
former.

Currently, this is only utilized by the `InstanceB` block,
but we would like to rework the other block constructors to also make
use of it.