Re #2829: allow instance arguments in pattern synonyms lhss...
... if they are instance arguments on the rhs as well.
To check this, we equip the scope checker with `Hiding` information
for `PatternBound` local variables, and propagate the hiding info from
`Arg` to the pattern scope checker by means of a new throw-away type
`WithHidingInfo` (to not clash with the existing `WithHiding`).

This commit achieves that instance arguments that come to surface
through pattern matching on a pattern synoym are available for
instance search, just if one would match against the pattern the
synonym stands for.