Refactor for #2829: Add MetaKind(InstanceMeta,UnificationMeta) to MetaInfo
We can now remember in the abstract syntax `Underscore` which kind of
meta we want to generate: a `UnificationMeta` (default) or an
`InstanceMeta`.  The latter is not used yet here but will be needed
for `{{_}}` arguments created by the expansion of pattern synonyms.

Having `data MetaKind(InstanceMeta,UnificationMeta)` we also refactor
`MetaInstantiation(Open,OpenInstance)` into a single parameterized
constructor `OpenMeta MetaKind`.  This helps condensing pattern
matching in several locations, where `Open` and `OpenInstance` behaved
the same.
I considered keeping `Open` and `OpenInstance` as pattern synonym but
it wasn't worth it.

There was already a type `MetaKind` for directing eta-expansion for
metas.  This is a more obscure use than specifying the solution
strategy for a meta, so I bumped this, calling it `MetaClass` now.
This type is local to a only couple of modules, so renaming seems ok.

Changing the order of arguments in `Rules.Term.checkUnderscore` and
friends seemed also natural.

Since `newInteractionMetaArg` was never used to create an instance
meta, I simplified its code.