Re #2829: Make instance arguments in pattern synonyms work in expressions.
If a pattern synonym is used in an expression, we need to make sure
that its instance arguments become instance metas.

Agda only turns an instance argument into an instance meta when the
argument is entirely omitted.  That is, `{{_}}` does not create an
instance meta, but this is exactly what the pattern synonym expansion
produces in expressions.

Having added `metaKind` to the `MetaInfo` record, we can now create
instance underscores in the abstract syntax, which fixes our dilemma.