New error warning `ConstructorDoesNotFitInData` instead of hard error.
When we check that the sort of the constructor (resp. constructor
argument) fits into the sort of the data type, we now catch the error
and wrap it into the `ConstructorDoesNotFitData` warning.

This warning is not benign and cannot be turned off,
but one can continue coding interactively now as this is no longer a
hard error.