Remove manual implementation of instance EmbPrj WarningName
Since `WarningName` is an Enum type, its serialisation can be obtained generically.

There has been a discussion to not automatically generate
serialisation as folks would forget to bump the interface version when
adding new constructors.
But note that `WarningName` tracks the constructors of `Warning`,
so there is still manual work implementing the serialisation of a new
warning, prompting to increase the interface version.