Re #7455: Setup.hs: catch when Agda did not produce all agdai files
The cause of #7455 is atm still unclear to me, but this patch adds a
check that all .agdai files we ask cabal to copy do actually exist.
If they do not, we skip the copy step.

Rationale: Unless the user does a sudo install, copying the interface files isn't
essential at the time of installation.